App context adicionado
