+++ app/src/main/AndroidManifest.xml
@@ -12,6 +12,7 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
+        android:name="com.example.fundatecheroes.App"
         android:theme="@style/Theme.FundatecHeroes"
         tools:targetApi="31">
         <activity
