+++ app/src/main/java/com/example/fundatecheroes/character/presentation/Character.kt
@@ -5,6 +5,6 @@ data class Character(
     val url: String,
     val descricao: String,
     val heroiVilao: String,
-    val idade: Number,
+    val idade: String,
     val aniversario: String
 )
