+++ app/src/main/java/com/example/fundatecheroes/character/view/NewCharacterActivity.kt
@@ -11,8 +11,6 @@ import androidx.activity.viewModels
 import com.example.fundatecheroes.character.presentation.NewCharacterViewModel
 import com.example.fundatecheroes.character.presentation.ViewState
 import com.example.fundatecheroes.databinding.ActivityNewCharacterBinding
-import com.squareup.moshi.Moshi
-import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 
 class NewCharacterActivity : AppCompatActivity(), OnItemSelectedListener {
 
@@ -21,12 +19,6 @@ class NewCharacterActivity : AppCompatActivity(), OnItemSelectedListener {
     private lateinit var binding: ActivityNewCharacterBinding
 
     private val viewModel: NewCharacterViewModel by viewModels()
-
-    private val moshi by lazy {
-        Moshi.Builder()
-            .add(KotlinJsonAdapterFactory())
-            .build()
-    }
 
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -38,15 +30,13 @@ class NewCharacterActivity : AppCompatActivity(), OnItemSelectedListener {
         configSaveButton()
 
 
-        val characterString = moshi.adapter(Character::class.java)
-            .toJson(character)
 
         viewModel.viewState.observe(this) { state ->
             when (state) {
                 is
                 ViewState.ShowErrorNull -> toastCamposNull()
                 ViewState.ShowErrorUrl -> toastUrlInvalida()
-                ViewState.ShowSuccess -> salvar()
+                is ViewState.ShowSuccess -> salvar()
             }
         }
     }
@@ -88,12 +78,8 @@ class NewCharacterActivity : AppCompatActivity(), OnItemSelectedListener {
         Toast.makeText(this, "campos não podem ser vazios!", Toast.LENGTH_LONG).show()
     }
 
-    private fun salvar(characterString: String?) {
+    private fun salvar() {
         Toast.makeText(this, "Heroi/Vilão salvo!", Toast.LENGTH_LONG).show()
-
-        val preferences = getSharedPreferences("bd", MODE_PRIVATE)
-
-        preferences.edit().putString("character", characterString).commit()
     }
 
     override fun onItemSelected(
